Renamed package database/transitiondb to database/transition
