+++ src/main/java/world/bentobox/bentobox/database/transition/Json2MySQLDatabase.java
@@ -1,19 +1,19 @@
-package world.bentobox.bentobox.database.transitiondb;
+package world.bentobox.bentobox.database.transition;
 
 import world.bentobox.bentobox.database.AbstractDatabaseHandler;
 import world.bentobox.bentobox.database.DatabaseSetup;
 import world.bentobox.bentobox.database.json.JSONDatabase;
-import world.bentobox.bentobox.database.yaml.YamlDatabase;
+import world.bentobox.bentobox.database.mysql.MySQLDatabase;
 
 /**
  * @author tastybento
  * @since 1.5.0
  */
-public class Yaml2JsonDatabase implements DatabaseSetup {
+public class Json2MySQLDatabase implements DatabaseSetup {
 
     @Override
     public <T> AbstractDatabaseHandler<T> getHandler(Class<T> type) {
-        return new TransitionDatabaseHandler<>(type, new YamlDatabase().getHandler(type), new JSONDatabase().getHandler(type));
+        return new TransitionDatabaseHandler<>(type, new JSONDatabase().getHandler(type), new MySQLDatabase().getHandler(type));
     }
 
 }
